+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [12.0.0](https://github.com/pipedrive/client-php/compare/11.1.1...12.0.0) (2025-06-17)
+
+### Removed
+- Removed `deals_summary` from `GET /v1/stages/:id` and `GET /v1/pipelines/:id`
+- Removed `totals_convert_currency` query parameter from `GET /v1/pipelines/:id`
+- Removed `everyone` query parameter from `GET /v1/stages/:id`
+
 ## [11.1.1](https://github.com/pipedrive/client-php/compare/11.1.0...11.1.1) (2025-06-09)
 
 ### Added